Extract symphonia_player into a tori-player lib
